ServerModel/requirements.txt
absl-py==0.13.0
astor==0.8.1
autopep8==1.5.7
bottle==0.12.19
cached-property==1.5.2
certifi==2021.5.30
Cython==0.29.23
gast==0.5.0
google-pasta==0.2.0
grpcio==1.38.1
h5py==3.3.0
importlib-metadata==4.6.1
jieba-fast==0.53
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.21.0
pip==21.1.3
protobuf==3.17.3
pycodestyle==2.7.0
PyYAML==5.4.1
scipy==1.7.0
sentencepiece==0.1.96
setuptools==57.1.0
six==1.16.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.61.2
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wheel==0.36.2
wrapt==1.12.1
zipp==3.5.0
